refactor: rename shutdowndialog.py, set ShutdownWarningDialog as a public class
