+++ =4.0.0
@@ -1,0 +1,26 @@
+Looking in indexes: https://pypi.tuna.tsinghua.edu.cn/simple
+Requirement already satisfied: pylint in /home/huaide/anaconda3/envs/backintime/lib/python3.9/site-packages (3.3.6)
+Requirement already satisfied: flake8 in /home/huaide/anaconda3/envs/backintime/lib/python3.9/site-packages (7.2.0)
+Requirement already satisfied: ruff in /home/huaide/anaconda3/envs/backintime/lib/python3.9/site-packages (0.11.7)
+Requirement already satisfied: codespell in /home/huaide/anaconda3/envs/backintime/lib/python3.9/site-packages (2.4.1)
+Requirement already satisfied: reuse in /home/huaide/anaconda3/envs/backintime/lib/python3.9/site-packages (5.0.2)
+Requirement already satisfied: astroid<=3.4.0.dev0,>=3.3.8 in /home/huaide/anaconda3/envs/backintime/lib/python3.9/site-packages (from pylint) (3.3.9)
+Requirement already satisfied: dill>=0.2 in /home/huaide/anaconda3/envs/backintime/lib/python3.9/site-packages (from pylint) (0.4.0)
+Requirement already satisfied: isort!=5.13,<7,>=4.2.5 in /home/huaide/anaconda3/envs/backintime/lib/python3.9/site-packages (from pylint) (6.0.1)
+Requirement already satisfied: mccabe<0.8,>=0.6 in /home/huaide/anaconda3/envs/backintime/lib/python3.9/site-packages (from pylint) (0.7.0)
+Requirement already satisfied: platformdirs>=2.2 in /home/huaide/anaconda3/envs/backintime/lib/python3.9/site-packages (from pylint) (4.3.7)
+Requirement already satisfied: tomli>=1.1 in /home/huaide/anaconda3/envs/backintime/lib/python3.9/site-packages (from pylint) (2.2.1)
+Requirement already satisfied: tomlkit>=0.10.1 in /home/huaide/anaconda3/envs/backintime/lib/python3.9/site-packages (from pylint) (0.13.2)
+Requirement already satisfied: typing-extensions>=3.10 in /home/huaide/anaconda3/envs/backintime/lib/python3.9/site-packages (from pylint) (4.13.2)
+Requirement already satisfied: pycodestyle<2.14.0,>=2.13.0 in /home/huaide/anaconda3/envs/backintime/lib/python3.9/site-packages (from flake8) (2.13.0)
+Requirement already satisfied: pyflakes<3.4.0,>=3.3.0 in /home/huaide/anaconda3/envs/backintime/lib/python3.9/site-packages (from flake8) (3.3.2)
+Requirement already satisfied: Jinja2>=3.0.0 in /home/huaide/anaconda3/envs/backintime/lib/python3.9/site-packages (from reuse) (3.1.6)
+Requirement already satisfied: attrs>=21.3 in /home/huaide/anaconda3/envs/backintime/lib/python3.9/site-packages (from reuse) (25.3.0)
+Requirement already satisfied: binaryornot>=0.4.4 in /home/huaide/anaconda3/envs/backintime/lib/python3.9/site-packages (from reuse) (0.4.4)
+Requirement already satisfied: boolean.py>=3.8 in /home/huaide/anaconda3/envs/backintime/lib/python3.9/site-packages (from reuse) (5.0)
+Requirement already satisfied: click>=8.0 in /home/huaide/anaconda3/envs/backintime/lib/python3.9/site-packages (from reuse) (8.1.8)
+Requirement already satisfied: license-expression>=1.0 in /home/huaide/anaconda3/envs/backintime/lib/python3.9/site-packages (from reuse) (30.4.1)
+Requirement already satisfied: python-debian!=0.1.45,!=0.1.46,!=0.1.47,>=0.1.34 in /home/huaide/anaconda3/envs/backintime/lib/python3.9/site-packages (from reuse) (1.0.1)
+Requirement already satisfied: chardet>=3.0.2 in /home/huaide/anaconda3/envs/backintime/lib/python3.9/site-packages (from binaryornot>=0.4.4->reuse) (5.2.0)
+Requirement already satisfied: MarkupSafe>=2.0 in /home/huaide/anaconda3/envs/backintime/lib/python3.9/site-packages (from Jinja2>=3.0.0->reuse) (3.0.2)
+Requirement already satisfied: charset-normalizer in /home/huaide/anaconda3/envs/backintime/lib/python3.9/site-packages (from python-debian!=0.1.45,!=0.1.46,!=0.1.47,>=0.1.34->reuse) (3.4.1)
